requirements.txt
hydra_core==0.11.3
hydra_colorlog==0.1.4
julius
numpy
pystoi==0.3.3
six
sounddevice==0.4.0
torchaudio
torch
soundfile
